\name{print.cats2ranks}
\alias{print.cats2ranks}
\title{Print the result of cats2ranks}
\description{Print the result of cats2ranks.}
\usage{
 \method{print}{cats2ranks}(x,...)
}
\arguments{
 \item{x}{The result of cats2ranks.}
 \item{...}{a dummy argument to keep S3 methods happy}
}
\details{
 Displays the names and mean ranks of the output of \samp{cats2ranks} in
 order of numerically ascending ranks.
}
\value{nil}
\author{Jim Lemon}
\seealso{\link{cats2ranks}}
\keyword{misc}
