\name{print.meanranks}
\alias{print.meanranks}
\title{Print the result of meanranks}
\description{Print the result of meanranks.}
\usage{
 \method{print}{meanranks}(x,...)
}
\arguments{
 \item{x}{The result of meanranks.}
 \item{...}{a dummy argument to keep S3 methods happy}
}
\details{
 Displays the names and mean ranks of the output of \samp{meanranks} in
 order of numerically ascending ranks.
}
\value{nil}
\author{Jim Lemon}
\seealso{\link{meanranks}}
\keyword{misc}
