\name{print.page.trend.test}
\alias{print.page.trend.test}
\title{prints the L statistic for Page's trend test}
\description{
 prints the obtained L statistic and the associated probability for the normal
 approximation if the sample size is sufficiently large
}
\usage{
 \method{print}{page.trend.test}(x,...)
}
\arguments{
 \item{x}{an object returned from \code{page.trend.test}}
 \item{...}{arguments to be passed to \code{print}}
}
\value{
 nil
}
\author{Jim Lemon}
\keyword{misc}
