\name{print.lwstat}
\alias{print.lwstat}
\title{Print the result of lwscreen}
\description{Print the result of lwscreen.}
\usage{
 \method{print}{lwstat}(x,...)
}
\arguments{
 \item{x}{The result of lwscreen.}
 \item{...}{a dummy argument to keep S3 methods happy}
}
\details{
 Displays the output of \code{lwscreen}.
}
\value{nil}
\author{Jim Lemon}
\seealso{\code{\link{lwscreen}}}
\keyword{misc}
